@@ -20,26 +20,34 @@
   },
   "homepage": "https://github.com/TRAININGIT-REACT-II/Practica-Final#readme",
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.11",
+    "@mui/material": "^5.11.13",
+    "@reduxjs/toolkit": "^1.9.3",
     "bcrypt": "^5.1.0",
-    "fastify": "^4.13.0",
+    "fastify": "^4.14.1",
     "lowdb": "^1.0.0",
     "nanoid": "^3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^5",
+    "redux": "^4.2.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.0",
-    "@babel/preset-env": "^7.14.0",
-    "@babel/preset-react": "^7.13.13",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "concurrently": "^6.0.2",
-    "css-loader": "^5.2.4",
-    "html-webpack-plugin": "^5.3.1",
-    "nodemon": "^2.0.7",
-    "style-loader": "^2.0.0",
-    "webpack": "^5.36.2",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2"
+    "@babel/core": "^7.21.0",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "babel-jest": "^29.5.0",
+    "babel-loader": "^9.1.2",
+    "concurrently": "^7.6.0",
+    "css-loader": "^6.7.3",
+    "html-webpack-plugin": "^5.5.0",
+    "nodemon": "^2.0.21",
+    "style-loader": "^3.3.2",
+    "webpack": "^5.76.2",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.13.1"
   }
 }
