@@ -1,29 +1,28 @@
 // Servidor de ejemplo para la práctica final del curso
-const fastify = require("fastify")({ logger: true });
-const bcrypt = require("bcrypt");
-const low = require("lowdb");
-const FileSync = require("lowdb/adapters/FileSync");
-const path = require("path");
-const { nanoid } = require("nanoid");
+const fastify = require('fastify')({ logger: true })
+const bcrypt = require('bcrypt')
+const low = require('lowdb')
+const FileSync = require('lowdb/adapters/FileSync')
+const path = require('path')
+const { nanoid } = require('nanoid')
 
 // Configuracion
-const DB_PATH =
-  process.env.DB_PATH || path.resolve(__dirname, "..", "data", "db.json");
+const DB_PATH = process.env.DB_PATH || path.resolve(__dirname, '..', 'data', 'db.json')
 const DB_DEFAULTS = {
   users: [],
   notes: [],
-};
-const SALT_ROUNDS_TO_ENCRYPT = 5;
-const USER_ID_LENGTH = 8;
-const NOTE_ID_LENGTH = 8;
-const TOKEN_LENGTH = 24;
-const API_TOKEN_HEADER = "api-token";
+}
+const SALT_ROUNDS_TO_ENCRYPT = 5
+const USER_ID_LENGTH = 8
+const NOTE_ID_LENGTH = 8
+const TOKEN_LENGTH = 24
+const API_TOKEN_HEADER = 'api-token'
 
 // Cargamos la base de datos (lowdb)
 // @ref https://github.com/typicode/lowdb
-const dbAdapter = new FileSync(DB_PATH);
-const db = low(dbAdapter);
-db.defaults(DB_DEFAULTS).write();
+const dbAdapter = new FileSync(DB_PATH)
+const db = low(dbAdapter)
+db.defaults(DB_DEFAULTS).write()
 
 /**
  * Configura Fastify para comprobar el token y obtener el usuario
@@ -32,43 +31,44 @@ db.defaults(DB_DEFAULTS).write();
 // Esta configuración es más conveniente hacerla a través de un
 // plugin, pero por simplicidad la haré directamente aquí.
 // @ref https://www.fastify.io/docs/latest/Decorators/#decorate-request
-fastify.decorateRequest("token", null);
-fastify.decorateRequest("user", null);
-fastify.decorateRequest("authenticated", false);
+fastify.decorateRequest('token', null)
+fastify.decorateRequest('user', null)
+fastify.decorateRequest('authenticated', false)
 
 // Agregamos un hook para cargar el usuario en cada petición
 // si el token está presente.
-fastify.addHook("preHandler", (request, _reply, done) => {
-  const token = request.headers[API_TOKEN_HEADER];
-  request.token = token;
+fastify.addHook('preHandler', (request, _reply, done) => {
+  const token = request.headers[API_TOKEN_HEADER]
+  request.token = token
 
   if (token != null) {
-    const user = db.get("users").find({ token }).value();
+    const user = db.get('users').find({ token }).value()
 
     if (user != null) {
-      request.user = user;
-      request.authenticated = true;
+      request.user = user
+      request.authenticated = true
     }
   }
 
-  done();
-});
+  done()
+})
 
 // Una función para compborar si el usuario existe.
 // En otro caso, devuelve un 401
 const checkAuthentication = (request, reply, done) => {
+  console.log(request.token)
   if (request.token == null) {
     reply.code(401).send({
-      error: "El token no está presente en la petición",
-    });
+      error: 'El token no está presente en la petición',
+    })
   } else if (!request.authenticated) {
     reply.code(401).send({
-      error: "No existe un usuario asociado al token",
-    });
+      error: 'No existe un usuario asociado al token',
+    })
   }
 
-  done();
-};
+  done()
+}
 
 // Peticiones
 
@@ -79,9 +79,9 @@ const checkAuthentication = (request, reply, done) => {
  * Devuelve un simple OK. Se puede utilizar para comprobar si el servidor
  * funciona.
  */
-fastify.get("/api", async () => {
-  return { status: "ok" };
-});
+fastify.get('/api', async () => {
+  return { status: 'ok' }
+})
 
 /**
  * POST /register
@@ -90,22 +90,22 @@ fastify.get("/api", async () => {
  * Registra un nuevo usuario en el sistema. Recibe los parámetros por
  * el cuerpo del mensaje
  */
-fastify.post("/api/register", (request, reply) => {
-  const { username, password } = request.body;
+fastify.post('/api/register', (request, reply) => {
+  const { username, password } = request.body
 
-  const userExists = db.get("users").find({ username }).value();
+  const userExists = db.get('users').find({ username }).value()
 
   if (userExists != null) {
     reply.code(400).send({
-      error: "El nombre de usuario ya existe",
-    });
+      error: 'El nombre de usuario ya existe',
+    })
   } else {
     // Encriptamos la contraseña
     bcrypt.hash(password, SALT_ROUNDS_TO_ENCRYPT, (err, hash) => {
       if (err) {
         return reply.code(500).send({
-          error: "Hubo un error al almacenar la contraseña",
-        });
+          error: 'Hubo un error al almacenar la contraseña',
+        })
       }
 
       const newUser = {
@@ -113,19 +113,19 @@ fastify.post("/api/register", (request, reply) => {
         username: username,
         password: hash,
         token: nanoid(TOKEN_LENGTH),
-      };
+      }
 
       // Almacenamos el usuario en la DB
-      db.get("users").push(newUser).write();
+      db.get('users').push(newUser).write()
 
       return reply.send({
         id: newUser.id,
         username,
         token: newUser.token,
-      });
-    });
+      })
+    })
   }
-});
+})
 
 /**
  * POST /login
@@ -133,22 +133,22 @@ fastify.post("/api/register", (request, reply) => {
  *
  * Inicia sesión devolviendo el token para las peticiones
  */
-fastify.post("/api/login", (request, reply) => {
-  const { username, password } = request.body;
+fastify.post('/api/login', (request, reply) => {
+  const { username, password } = request.body
 
-  const user = db.get("users").find({ username }).value();
+  const user = db.get('users').find({ username }).value()
 
   if (user == null) {
     reply.code(401).send({
-      error: "Las credenciales introducidas no son correctas",
-    });
+      error: 'Las credenciales introducidas no son correctas',
+    })
   } else {
     // Encriptamos la contraseña
     bcrypt.compare(password, user.password, (err, result) => {
       if (err) {
         return reply.code(500).send({
-          error: "Hubo un error al comprobar las credenciales",
-        });
+          error: 'Hubo un error al comprobar las credenciales',
+        })
       }
 
       if (result) {
@@ -156,15 +156,15 @@ fastify.post("/api/login", (request, reply) => {
           id: user.id,
           username,
           token: user.token,
-        });
+        })
       } else {
         reply.code(401).send({
-          error: "Las credenciales introducidas no son correctas",
-        });
+          error: 'Las credenciales introducidas no son correctas',
+        })
       }
-    });
+    })
   }
-});
+})
 
 /**
  * GET /notes
@@ -173,17 +173,17 @@ fastify.post("/api/login", (request, reply) => {
  * Devuelve las notas del usuario
  */
 fastify.route({
-  method: "GET",
-  path: "/api/notes",
+  method: 'GET',
+  path: '/api/notes',
   preHandler: checkAuthentication,
   handler: (request, reply) => {
     // Obtenemos el usuario
-    const { id: userId } = request.user;
-    const notes = db.get("notes").filter({ author: userId }).value();
+    const { id: userId } = request.user
+    const notes = db.get('notes').filter({ author: userId }).value()
 
-    return notes;
+    return notes
   },
-});
+})
 
 /**
  * GET /notes/:id
@@ -192,19 +192,19 @@ fastify.route({
  * Devuelve las notas del usuario
  */
 fastify.route({
-  method: "GET",
-  path: "/api/notes/:id",
+  method: 'GET',
+  path: '/api/notes/:id',
   preHandler: checkAuthentication,
   handler: (request, reply) => {
     // Obtenemos el usuario
-    const { id: userId } = request.user;
-    const { id } = request.params;
-    const note = db.get("notes").find({ author: userId, id }).value();
+    const { id: userId } = request.user
+    const { id } = request.params
+    const note = db.get('notes').find({ author: userId, id }).value()
 
     if (note == null) {
       reply.code(404).send({
-        error: "La nota no existe",
-      });
+        error: 'La nota no existe',
+      })
     } else {
       return {
         ...note,
@@ -212,10 +212,10 @@ fastify.route({
           id: userId,
           username: request.user.username,
         },
-      };
+      }
     }
   },
-});
+})
 
 /**
  * POST /notes
@@ -224,14 +224,14 @@ fastify.route({
  * Crea una nueva nota en el sistema
  */
 fastify.route({
-  method: "POST",
-  path: "/api/notes",
+  method: 'POST',
+  path: '/api/notes',
   preHandler: checkAuthentication,
   handler: (request, reply) => {
     // Obtenemos el usuario
-    const { id: userId } = request.user;
-    const { title, content } = request.body;
-    const date = new Date();
+    const { id: userId } = request.user
+    const { title, content } = request.body
+    const date = new Date()
 
     const newPost = {
       author: userId,
@@ -240,19 +240,19 @@ fastify.route({
       createdAt: date,
       updatedAt: date,
       id: nanoid(NOTE_ID_LENGTH),
-    };
+    }
 
-    db.get("notes").push(newPost).write();
+    db.get('notes').push(newPost).write()
 
     return {
       id: newPost.id,
       title,
       content,
       createdAt: date,
-      updatedAt: date
-    };
+      updatedAt: date,
+    }
   },
-});
+})
 
 /**
  * PATCH /notes/id
@@ -261,26 +261,23 @@ fastify.route({
  * Actualiza una nota en el sistema
  */
 fastify.route({
-  method: ["PATCH", "PUT"],
-  path: "/api/notes/:id",
+  method: ['PATCH', 'PUT'],
+  path: '/api/notes/:id',
   preHandler: checkAuthentication,
   handler: (request, reply) => {
     // Obtenemos el usuario
-    const { id: userId } = request.user;
-    const { title, content } = request.body;
-    const { id } = request.params;
-    const note = db.get("notes").find({ author: userId, id }).value();
+    const { id: userId } = request.user
+    const { title, content } = request.body
+    const { id } = request.params
+    const note = db.get('notes').find({ author: userId, id }).value()
 
     if (note == null) {
       reply.code(404).send({
-        error: "La nota no existe",
-      });
+        error: 'La nota no existe',
+      })
     } else {
       // Actualizamos
-      db.get("notes")
-        .find({ author: userId, id })
-        .assign({ title, content, updatedAt: new Date() })
-        .write();
+      db.get('notes').find({ author: userId, id }).assign({ title, content, updatedAt: new Date() }).write()
 
       return {
         ...note,
@@ -290,10 +287,10 @@ fastify.route({
           id: userId,
           username: request.user.username,
         },
-      };
+      }
     }
   },
-});
+})
 
 /**
  * DELETE /notes/id
@@ -302,38 +299,38 @@ fastify.route({
  * Crea una nueva nota en el sistema
  */
 fastify.route({
-  method: "DELETE",
-  path: "/api/notes/:id",
+  method: 'DELETE',
+  path: '/api/notes/:id',
   preHandler: checkAuthentication,
   handler: (request, reply) => {
     // Obtenemos el usuario
-    const { id: userId } = request.user;
-    const { id } = request.params;
-    const note = db.get("notes").find({ author: userId, id }).value();
+    const { id: userId } = request.user
+    const { id } = request.params
+    const note = db.get('notes').find({ author: userId, id }).value()
 
     if (note == null) {
       reply.code(404).send({
-        error: "La nota no existe",
-      });
+        error: 'La nota no existe',
+      })
     } else {
       // Eliminamos
-      db.get("notes").remove({ author: userId, id }).write();
+      db.get('notes').remove({ author: userId, id }).write()
 
-      return {};
+      return {}
     }
   },
-});
+})
 
 // Iniciamos el servidor!
 const start = async () => {
   try {
     await fastify.listen({
-      port: 3000
-    });
+      port: 3000,
+    })
   } catch (err) {
-    fastify.log.error(err);
-    process.exit(1);
+    fastify.log.error(err)
+    process.exit(1)
   }
-};
+}
 
-start();
+start()
